LayoutTests/webaudio/audiocontext-promise-expected.txt

Tests that AudioContext suspend method resolve promises immediately when already suspended.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.state is "suspended"
PASS context.state is "suspended"
PASS context resolve callback called
PASS context.state is "suspended"
PASS successfullyParsed is true

TEST COMPLETE
